Add graphite_web_config resource create action
The graphite_web_config resource is a generic resource for writing a
properly formatted file such as a local_settings.py needed by django.
It uses the ChefGraphite::PythonWriter class. It is not limited to
managing the local_settings.py and will manage any file.

Includes ChefSpec tests for the resource and provider and functional
example recipe for integration test